refactor: 디렉터리 구조 변경 및 ticket.js 추가
