+++ src/js/models/lotto_model.js
@@ -1,4 +1,4 @@
-import { INIT } from "./constant.js"
+import { INIT } from "../constants/constant.js"
 
 class LottoModel {
   constructor() {
