+++ src/js/templates/lotto_template.js
